+++ googlemaps/src/main/java/com/vaadin/tapio/googlemaps/client/events/MapMoveListener.java
@@ -22,6 +22,6 @@ public interface MapMoveListener extends Serializable {
      * @param boundsSW
      *            The position of the south-west corner of the map.
      */
-    public void mapMoved(double zoomLevel, LatLon center, LatLon boundsNE,
-            LatLon boundsSW);
+    public void mapMoved(int zoomLevel, LatLon center, LatLon boundsNE,
+        LatLon boundsSW);
 }
